pbd/pipelines/fast_inference/pyproject.toml
[project]
name = "fast-inference"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "mistralrs-cuda>=0.5.0",
    "pillow>=11.2.1",
]
